♻️ refacotr: split remaining __main__ and unittests
